# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrackFlow API — digital tracking governance platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/dashboard:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard KPI statistics
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: campaignId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /workspaces/{workspaceId}/dashboard/timeline:
    get:
      operationId: getDashboardTimeline
      tags: [dashboard]
      summary: Get click/conversion timeline for chart
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: granularity
          in: query
          schema:
            type: string
            enum: [day, week, month]
      responses:
        "200":
          description: Timeline data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelinePoint"

  # ─── Campaigns ────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/links:
    get:
      operationId: listLinks
      tags: [links]
      summary: List all tracked links in a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: campaignId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackedLink"
    post:
      operationId: createLink
      tags: [links]
      summary: Create a new tracked link
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedLink"
  /workspaces/{workspaceId}/links/{linkId}:
    get:
      operationId: getLink
      tags: [links]
      summary: Get a link by ID
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: linkId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedLink"
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a tracked link
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: linkId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workspaces/{workspaceId}/links/generate-utm:
    post:
      operationId: generateUtm
      tags: [links]
      summary: Generate UTM parameters from campaign info
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UtmGeneratorInput"
      responses:
        "200":
          description: Generated UTM parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UtmResult"
  /workspaces/{workspaceId}/links/validate:
    post:
      operationId: validateLink
      tags: [links]
      summary: Validate a URL for UTM correctness
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkValidationInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkValidationResult"

  # ─── Webhooks ─────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all tracking events (click, lead, webhook, sale)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
        - name: leadId
          in: query
          schema:
            type: ["string", "null"]
        - name: eventType
          in: query
          schema:
            type: ["string", "null"]
        - name: campaignId
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackingEvent"

  # ─── Conversions ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DashboardStats:
      type: object
      required: [clicks, leads, conversions, revenue, roi, cpl, cpa, roas, topCampaigns]
      properties:
        clicks:
          type: integer
        leads:
          type: integer
        conversions:
          type: integer
        revenue:
          type: number
        roi:
          type: number
        cpl:
          type: number
        cpa:
          type: number
        roas:
          type: number
        topCampaigns:
          type: array
          items:
            $ref: "#/components/schemas/CampaignStats"
    TimelinePoint:
      type: object
      required: [date, clicks, leads, conversions, revenue]
      properties:
        date:
          type: string
        clicks:
          type: integer
        leads:
          type: integer
        conversions:
          type: integer
        revenue:
          type: number

    # ─── Campaign ──────────────────────────────────────────────────────────
...
    CampaignStats:
      type: object
      required: [campaignId, campaignName, clicks, leads, conversions, revenue, roas]
      properties:
        campaignId:
          type: integer
        campaignName:
          type: string
        clicks:
          type: integer
        leads:
          type: integer
        conversions:
          type: integer
        revenue:
          type: number
        roas:
          type: number

    # ─── Links ─────────────────────────────────────────────────────────────
    TrackedLink:
      type: object
      required: [id, workspaceId, originalUrl, finalUrl, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        campaignId:
          type: ["integer", "null"]
        campaignName:
          type: ["string", "null"]
        originalUrl:
          type: string
        finalUrl:
          type: string
        shortUrl:
          type: ["string", "null"]
        qrCodeUrl:
          type: ["string", "null"]
        utmSource:
          type: ["string", "null"]
        utmMedium:
          type: ["string", "null"]
        utmCampaign:
          type: ["string", "null"]
        utmContent:
          type: ["string", "null"]
        utmTerm:
          type: ["string", "null"]
        responsible:
          type: ["string", "null"]
        clicks:
          type: integer
        createdAt:
          type: string
    LinkInput:
      type: object
      required: [originalUrl]
      properties:
        originalUrl:
          type: string
        campaignId:
          type: integer
        utmSource:
          type: string
        utmMedium:
          type: string
        utmCampaign:
          type: string
        utmContent:
          type: string
        utmTerm:
          type: string
        responsible:
          type: string
    UtmGeneratorInput:
      type: object
      required: [baseUrl, platform, campaignName]
      properties:
        baseUrl:
          type: string
        platform:
          type: string
        campaignName:
          type: string
        creative:
          type: string
        audience:
          type: string
        campaignId:
          type: integer
    UtmResult:
      type: object
      required: [finalUrl, utmSource, utmMedium, utmCampaign]
      properties:
        finalUrl:
          type: string
        utmSource:
          type: string
        utmMedium:
          type: string
        utmCampaign:
          type: string
        utmContent:
          type: ["string", "null"]
        utmTerm:
          type: ["string", "null"]
    LinkValidationInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
    LinkValidationResult:
      type: object
      required: [valid, issues, suggestions]
      properties:
        valid:
          type: boolean
        issues:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        parsedParams:
          type: object
          additionalProperties:
            type: string

    # ─── Webhooks ──────────────────────────────────────────────────────────
...
    TrackingEvent:
      type: object
      required: [id, workspaceId, eventType, timestamp]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        eventType:
          type: string
          enum: [click, lead, webhook, crm, sale]
        leadId:
          type: ["string", "null"]
        campaignId:
          type: ["integer", "null"]
        campaignName:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        medium:
          type: ["string", "null"]
        metadata:
          type: object
          additionalProperties: true
        timestamp:
          type: string

    # ─── Conversions ───────────────────────────────────────────────────────
...